Check the users input
Get the input and convert it to uppercase and then add it to the current node as its answer, later iterating through the nodes and checking to see if each answer is correct or not.
Finally see if the user has completed the grid or not and move to the next node / reset the textbox text.